refactor: move composio catalog tests to dedicated module
